bump version to 0.21.0-rc.1 (#1496)
