+++ .github/workflows/main.yml
@@ -421,4 +421,5 @@ jobs:
       - run: scripts/install_prover.sh
         working-directory: compact_proof
       - run: cargo test -p risc0-groth16 -F docker --release
+      - run: cargo test -p risc0-zkvm -F prove -F docker -- tests::stark2snark
       - run: sccache --show-stats
