+++ .gitignore
@@ -15,3 +15,4 @@ target/
 tmp/
 # Ignore .pb files generated by profiling
 *.pb
+tools/smoke-test/Cargo.lock
