+++ Cargo.toml
@@ -25,31 +25,31 @@ members = [
 exclude = ["tools/crates-validator"]
 
 [workspace.package]
-version = "0.21.0-alpha.1"
+version = "0.21.0-rc.1"
 edition = "2021"
 license = "Apache-2.0"
 homepage = "https://risczero.com/"
 repository = "https://github.com/risc0/risc0/"
 
 [workspace.dependencies]
-bonsai-ethereum-contracts = { version = "0.6.0-rc.1", path = "bonsai/ethereum" }
-bonsai-ethereum-relay = { version = "0.6.0-rc.1", default-features = false, path = "bonsai/ethereum-relay" }
-bonsai-rest-api-mock = { version = "0.6.0-rc.1", default-features = false, path = "bonsai/rest-api-mock" }
-bonsai-sdk = { version = "0.6.0-rc.1", default-features = false, path = "bonsai/sdk" }
-risc0-binfmt = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/binfmt" }
-risc0-build = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/build" }
-risc0-build-kernel = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/build_kernel" }
-risc0-circuit-recursion = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/circuit/recursion" }
-risc0-circuit-recursion-sys = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/circuit/recursion-sys" }
-risc0-circuit-rv32im = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/circuit/rv32im" }
-risc0-circuit-rv32im-sys = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/circuit/rv32im-sys" }
-risc0-core = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/core" }
-risc0-groth16 = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/groth16" }
-risc0-r0vm = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/r0vm" }
-risc0-sys = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/sys" }
-risc0-zkp = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/zkp" }
-risc0-zkvm = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/zkvm" }
-risc0-zkvm-platform = { version = "0.21.0-alpha.1", default-features = false, path = "risc0/zkvm/platform" }
+bonsai-ethereum-contracts = { version = "0.7.0-rc.1", path = "bonsai/ethereum" }
+bonsai-ethereum-relay = { version = "0.7.0-rc.1", default-features = false, path = "bonsai/ethereum-relay" }
+bonsai-rest-api-mock = { version = "0.7.0-rc.1", default-features = false, path = "bonsai/rest-api-mock" }
+bonsai-sdk = { version = "0.7.0-rc.1", default-features = false, path = "bonsai/sdk" }
+risc0-binfmt = { version = "0.21.0-rc.1", default-features = false, path = "risc0/binfmt" }
+risc0-build = { version = "0.21.0-rc.1", default-features = false, path = "risc0/build" }
+risc0-build-kernel = { version = "0.21.0-rc.1", default-features = false, path = "risc0/build_kernel" }
+risc0-circuit-recursion = { version = "0.21.0-rc.1", default-features = false, path = "risc0/circuit/recursion" }
+risc0-circuit-recursion-sys = { version = "0.21.0-rc.1", default-features = false, path = "risc0/circuit/recursion-sys" }
+risc0-circuit-rv32im = { version = "0.21.0-rc.1", default-features = false, path = "risc0/circuit/rv32im" }
+risc0-circuit-rv32im-sys = { version = "0.21.0-rc.1", default-features = false, path = "risc0/circuit/rv32im-sys" }
+risc0-core = { version = "0.21.0-rc.1", default-features = false, path = "risc0/core" }
+risc0-groth16 = { version = "0.21.0-rc.1", default-features = false, path = "risc0/groth16" }
+risc0-r0vm = { version = "0.21.0-rc.1", default-features = false, path = "risc0/r0vm" }
+risc0-sys = { version = "0.21.0-rc.1", default-features = false, path = "risc0/sys" }
+risc0-zkp = { version = "0.21.0-rc.1", default-features = false, path = "risc0/zkp" }
+risc0-zkvm = { version = "0.21.0-rc.1", default-features = false, path = "risc0/zkvm" }
+risc0-zkvm-platform = { version = "0.21.0-rc.1", default-features = false, path = "risc0/zkvm/platform" }
 
 [profile.bench]
 lto = true
