+++ bonsai/Cargo.toml
@@ -9,12 +9,12 @@ homepage = "https://risczero.com/"
 repository = "https://github.com/risc0/risc0/"
 
 [workspace.dependencies]
-bonsai-ethereum-contracts = { version = "0.6.0-rc.1", path = "ethereum" }
-bonsai-ethereum-relay = { version = "0.6.0-rc.1", default-features = false, path = "ethereum-relay" }
-bonsai-rest-api-mock = { version = "0.6.0-rc.1", default-features = false, path = "rest-api-mock" }
-bonsai-sdk = { version = "0.6.0-rc.1", default-features = false, path = "sdk" }
-risc0-build = { version = "0.21.0-alpha.1", default-features = false, path = "../risc0/build" }
-risc0-zkvm = { version = "0.21.0-alpha.1", default-features = false, path = "../risc0/zkvm" }
+bonsai-ethereum-contracts = { version = "0.7.0-rc.1", path = "ethereum" }
+bonsai-ethereum-relay = { version = "0.7.0-rc.1", default-features = false, path = "ethereum-relay" }
+bonsai-rest-api-mock = { version = "0.7.0-rc.1", default-features = false, path = "rest-api-mock" }
+bonsai-sdk = { version = "0.7.0-rc.1", default-features = false, path = "sdk" }
+risc0-build = { version = "0.21.0-rc.1", default-features = false, path = "../risc0/build" }
+risc0-zkvm = { version = "0.21.0-rc.1", default-features = false, path = "../risc0/zkvm" }
 
 [profile.bench]
 lto = true
