+++ bonsai/ethereum-relay/Cargo.toml
@@ -1,11 +1,14 @@
 [package]
 name = "bonsai-ethereum-relay"
 description = "A relayer to integrate Ethereum with Bonsai."
-version = "0.6.0-rc.1"
+version = "0.7.0-rc.1"
 edition = { workspace = true }
 license = { workspace = true }
 homepage = { workspace = true }
 repository = { workspace = true }
+
+[package.metadata.release]
+release = false
 
 [dependencies]
 anyhow = "1.0"
