+++ bonsai/rest-api-mock/Cargo.toml
@@ -1,11 +1,14 @@
 [package]
 name = "bonsai-rest-api-mock"
 description = "An HTTP REST API server to mock the Bonsai-alpha prover interface."
-version = "0.6.0-rc.1"
+version = "0.7.0-rc.1"
 edition = { workspace = true }
 license = { workspace = true }
 homepage = { workspace = true }
 repository = { workspace = true }
+
+[package.metadata.release]
+release = false
 
 [dependencies]
 anyhow = "1.0"
