+++ bonsai/sdk/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "bonsai-sdk"
 description = "Bonsai Software Development Kit"
-version = "0.6.0-rc.1"
+version = "0.7.0-rc.1"
 edition = { workspace = true }
 license = { workspace = true }
 homepage = { workspace = true }
@@ -22,7 +22,7 @@ anyhow = "1.0"
 bincode = "1.3"
 bytemuck = "1.14"
 hex = "0.4"
-httpmock = "0.6"  
+httpmock = "0.6"
 risc0-zkvm = { path = "../../risc0/zkvm" }
 temp-env = "0.3"
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
