+++ risc0/zkvm/src/host/server/prove/tests.rs
@@ -456,6 +456,53 @@ mod docker {
     }
 }
 
+#[cfg(feature = "docker")]
+#[test]
+fn stark2snark() {
+    use crate::{
+        get_prover_server, recursion::identity_p254, CompactReceipt, ExecutorEnv, ExecutorImpl,
+        InnerReceipt, ProverOpts, Receipt, VerifierContext,
+    };
+    use risc0_groth16::docker::stark_to_snark;
+    use risc0_zkvm_methods::{multi_test::MultiTestSpec, MULTI_TEST_ELF, MULTI_TEST_ID};
+
+    let env = ExecutorEnv::builder()
+        .write(&MultiTestSpec::BusyLoop { cycles: 0 })
+        .unwrap()
+        .build()
+        .unwrap();
+
+    tracing::info!("execute");
+
+    let mut exec = ExecutorImpl::from_elf(env, MULTI_TEST_ELF).unwrap();
+    let session = exec.run().unwrap();
+
+    tracing::info!("prove");
+    let opts = ProverOpts::default();
+    let ctx = VerifierContext::default();
+    let prover = get_prover_server(&opts).unwrap();
+    let receipt = prover.prove_session(&ctx, &session).unwrap();
+    let claim = receipt.get_claim().unwrap();
+    let composite_receipt = receipt.inner.composite().unwrap();
+    let succinct_receipt = prover.compress(composite_receipt).unwrap();
+    let journal = session.journal.unwrap().bytes;
+
+    tracing::info!("identity_p254");
+    let ident_receipt = identity_p254(&succinct_receipt).unwrap();
+    let seal_bytes = ident_receipt.get_seal_bytes();
+
+    tracing::info!("stark-to-snark");
+    let seal = stark_to_snark(&seal_bytes).unwrap().to_vec();
+
+    tracing::info!("Receipt");
+    let receipt = Receipt::new(
+        InnerReceipt::Compact(CompactReceipt { seal, claim }),
+        journal,
+    );
+
+    receipt.verify(MULTI_TEST_ID).unwrap();
+}
+
 mod sys_verify {
     use risc0_zkvm_methods::{
         multi_test::MultiTestSpec, HELLO_COMMIT_ELF, HELLO_COMMIT_ID, MULTI_TEST_ELF, MULTI_TEST_ID,
