+++ tools/smoke-test/methods/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [build-dependencies]
-risc0-build = "0.21.0-alpha.1"
+risc0-build = "0.21.0-rc.1"
 # Use this only for testing locally before a release.
 # risc0-build = { path = "../../../risc0/build" }
